+++ src/algo_royale/services/ledger_service.py
@@ -86,7 +86,7 @@ class LedgerService:
         self,
         order_id: str,
         status: str,
-        quantity: int | None = None,
+        quantity: float | None = None,
         price: float | None = None,
     ) -> None:
         """Update an existing order."""
